requirements.txt
py-cord==2.4.1
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0